build(gulp): refactor gulpfile into tasks, fixes #3823
