+++ gulp-tasks/copy-fonts.js
@@ -1,0 +1,10 @@
+const gulp = require('gulp');
+
+/**
+ * @returns {NodeJS.ReadWriteStream}
+ */
+const copyFonts = () => {
+  return gulp.src('src/fonts/**/*').pipe(gulp.dest('dist/fonts/'));
+};
+
+module.exports = copyFonts;
