+++ gulpfile.js
@@ -15,147 +15,35 @@
  */
 
 const gulp = require('gulp');
+const copyContentAssets = require('./gulp-tasks/copy-content-assets.js');
+const copyDefaultLocale = require('./gulp-tasks/copy-default-locale.js');
+const copyFonts = require('./gulp-tasks/copy-fonts.js');
+const copyGlobalImages = require('./gulp-tasks/copy-global-images.js');
+const copyMisc = require('./gulp-tasks/copy-misc.js');
 const sassTask = require('./gulp-tasks/sass.js');
-const mozjpeg = require('imagemin-mozjpeg');
-const pngquant = require('imagemin-pngquant');
-const imagemin = require('gulp-imagemin');
-const rename = require('gulp-rename');
-const through2 = require('through2');
-
-const {defaultLocale} = require('./src/site/_data/site');
-
-/* eslint-disable max-len */
-const assetTypes =
-  'jpg,jpeg,png,svg,gif,webp,webm,mp4,mov,ogg,wav,mp3,txt,yaml';
-/* eslint-enable max-len */
-
-const isProd = process.env.ELEVENTY_ENV === 'prod';
-
-const compressImagesTransform = (pngQuality, jpegQuality) => {
-  if (!isProd) {
-    // This is the identity transform, which does nothing and just passes
-    // the images through directly.
-    return through2.obj();
-  }
-  return imagemin([
-    // @ts-ignore
-    pngquant({quality: [pngQuality, pngQuality]}),
-    mozjpeg({quality: jpegQuality * 100}),
-  ]);
-};
-
-// These are global images used in CSS and base HTML, such as author profiles.
-// We don't compress the PNGs here as they are trivially small (site icons).
-gulp.task('copy-global-images', () => {
-  return gulp
-    .src(['./src/images/**/*'])
-    .pipe(compressImagesTransform(1.0, 0.8))
-    .pipe(gulp.dest('./dist/images'));
-});
-
-// These are misc top-level assets.
-gulp.task('copy-misc', () => {
-  return gulp.src(['./src/misc/**/*']).pipe(gulp.dest('./dist'));
-});
-
-// Images and any other assets in the content directory that should be copied
-// over along with the posts themselves.
-// Because we use permalinks to strip the parent directories from our posts
-// we need to also strip them from the content paths.
-gulp.task('copy-content-assets', () => {
-  return (
-    gulp
-      .src([`./src/site/content/**/*.{${assetTypes}}`])
-      .pipe(compressImagesTransform(0.8, 0.8))
-      // This makes the images show up in the same spot as the permalinked posts
-      // they belong to.
-      .pipe(
-        rename((assetPath) => {
-          const defaultLocaleRegExp = new RegExp(`^${defaultLocale}/`);
-          const parts = assetPath.dirname.split('/');
-          assetPath.dirname = assetPath.dirname.replace(
-            defaultLocaleRegExp,
-            '',
-          );
-          // Landing pages should keep their assets.
-          // e.g. en/vitals, en/about
-          if (parts.length <= 2) {
-            return;
-          }
-
-          // Let images pass through if they're special, i.e. part of the
-          // handbook or newsletter. We don't treat these URLs like the
-          // rest of the site and they're allowed to nest.
-          const subdir = parts[1];
-          const imagePassThroughs = ['handbook', 'newsletter'];
-          if (imagePassThroughs.includes(subdir)) {
-            return;
-          }
-
-          // Some assets are nested under directories which aren't part of
-          // their url. For example, we have /en/blog/some-post/foo.jpg.
-          // For these assets we need to remove the /blog/ directory so they
-          // can live at /some-post/foo.jpg since that's what we'll actually
-          // serve in production.
-          // e.g. en/blog/foo/bar.jpg -> /foo/bar.jpg
-          parts.splice(1, 1);
-          assetPath.dirname = parts.join('/').replace(defaultLocaleRegExp, '');
-        }),
-      )
-      .pipe(gulp.dest('./dist/'))
-  );
-});
-
-gulp.task('copy-fonts', () => {
-  return gulp.src('src/fonts/**/*').pipe(gulp.dest('dist/fonts/'));
-});
-
-/**
- * Copies content of the default locale (en) into its own folder
- * on prod builds.
- *
- * This is done for when we deploy to Firebase we need an `i18n/en`
- * folder for users who have English as their primary language,
- * and a second language that is supported by us. Firebase hosting
- * will serve the secondary language as it tries to host i81n content first.
- */
-gulp.task('default-locale', () => {
-  if (isProd) {
-    return gulp
-      .src(['./dist/**/*.html', '!./dist/i18n/**/*.html'])
-      .pipe(gulp.dest(`./dist/i18n/${defaultLocale}`));
-  }
-  return Promise.resolve();
-});
 
 gulp.task('sass', sassTask);
+
+gulp.task('default-locale', copyDefaultLocale);
 
 gulp.task(
   'build',
   gulp.parallel(
-    'copy-global-images',
-    'copy-misc',
-    'copy-content-assets',
-    'copy-fonts',
+    copyGlobalImages,
+    copyMisc,
+    copyContentAssets,
+    copyFonts,
     sassTask,
   ),
 );
 
 gulp.task('watch', () => {
-  gulp.watch(
-    './src/images/**/*',
-    {ignoreInitial: true},
-    gulp.series('copy-global-images'),
-  );
-  gulp.watch(
-    './src/misc/**/*',
-    {ignoreInitial: true},
-    gulp.series('copy-misc'),
-  );
+  gulp.watch('./src/images/**/*', {ignoreInitial: true}, copyGlobalImages);
+  gulp.watch('./src/misc/**/*', {ignoreInitial: true}, copyMisc);
   gulp.watch(
     './src/site/content/**/*',
     {ignoreInitial: true},
-    gulp.series('copy-content-assets'),
+    copyContentAssets,
   );
   gulp.watch('./src/styles/**/*.scss', {ignoreInitial: true}, sassTask);
 });
