+++ package.json
@@ -77,6 +77,7 @@
     "@rollup/plugin-node-resolve": "^11.1.1",
     "@rollup/plugin-virtual": "^2.0.3",
     "@types/google.analytics": "0.0.40",
+    "@types/gulp": "^4.0.8",
     "@types/node": "^13.11.1",
     "@types/remove-markdown": "^0.1.1",
     "@types/resize-observer-browser": "^0.1.3",
